refactor(direct-control): extract GameInfo row types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only GameInfo row input/result type ownership from packages/civ7-direct-control/src/index.ts into packages/civ7-direct-control/src/play/map/gameinfo.ts while keeping facade type re-exports in index.ts.

Preserves public type contracts and does not change GameInfo table or filter validation, lookup/filter semantics, source strings, runtime behavior, AI ingestion, static profile shaping, procedure schemas, telemetry, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- map-and-visibility.test.ts runtime-and-catalog.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.control.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, hotseat product-path, AI-ingestion, static profile, semantic CLI, telemetry, or procedure-core proof claimed.
