+++ openspec/changes/civ7-support-direct-control-modularization/workstream/direct-control-atom-corpus.md
@@ -164,6 +164,14 @@ schemas, telemetry, AI ingestion, hotseat runtime proof, CLI semantic
 projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row
 acceptance.
 
+GameInfo row type update: `src/play/map/gameinfo.ts` now also owns GameInfo row
+input/result types while public facade type re-exports stay in `index.ts`. This
+is local package/type relocation proof only and does not change GameInfo table
+or filter validation, lookup/filter semantics, source strings, runtime behavior,
+AI ingestion, static profile shaping, procedure schemas, telemetry, hotseat
+runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task
+2.9.4 matrix-row acceptance.
+
 Bounded root inspection update: `src/runtime/root-inspection.ts` now owns
 `inspectCiv7Root` and the generated bounded root inspection command while the
 public facade export stays in `index.ts` and injects command execution,
