+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -639,6 +639,17 @@ All future agent waves must be framed before delegation:
   2.9.4 matrix-row acceptance. This is local package/type relocation proof
   only, not runtime proof, AI/hotseat product-path support, or procedure-core
   readiness.
+- Direct-control GameInfo row types slice: completed as a narrow public type
+  ownership relocation. It moves only GameInfo row input/result types into
+  `src/play/map/gameinfo.ts` while keeping facade type re-exports in `index.ts`
+  and replacing the GameInfo module's type-only facade imports with concrete
+  owner imports. This preserves public type contracts and does not change
+  GameInfo table or filter validation, lookup/filter semantics, source strings,
+  runtime behavior, AI ingestion, static profile shaping, procedure schemas,
+  telemetry, hotseat runtime proof, CLI semantic projection, Effect/oRPC
+  procedure-core work, or Task 2.9.4 matrix-row acceptance. This is local
+  package/type relocation proof only, not runtime proof, AI/hotseat product-path
+  support, or procedure-core readiness.
 - Direct-control turn-completion slice: completed as a narrow turn-completion
   source/wrapper relocation. It moves only `getCiv7TurnCompletionStatus`,
   `sendCiv7TurnComplete`, `sendCiv7TurnUnready`, the status command builder,
